+++ vs-20181212/src/Models/ListRenderingImagesResponseBody.php
@@ -1,0 +1,122 @@
+<?php
+
+// This file is auto-generated, don't edit it. Thanks.
+
+namespace AlibabaCloud\SDK\Vs\V20181212\Models;
+
+use AlibabaCloud\Dara\Model;
+use AlibabaCloud\SDK\Vs\V20181212\Models\ListRenderingImagesResponseBody\images;
+
+class ListRenderingImagesResponseBody extends Model
+{
+    /**
+     * @var images[]
+     */
+    public $images;
+
+    /**
+     * @var string
+     */
+    public $pageNumber;
+
+    /**
+     * @var string
+     */
+    public $pageSize;
+
+    /**
+     * @var string
+     */
+    public $requestId;
+
+    /**
+     * @var int
+     */
+    public $totalCount;
+    protected $_name = [
+        'images' => 'Images',
+        'pageNumber' => 'PageNumber',
+        'pageSize' => 'PageSize',
+        'requestId' => 'RequestId',
+        'totalCount' => 'TotalCount',
+    ];
+
+    public function validate()
+    {
+        if (\is_array($this->images)) {
+            Model::validateArray($this->images);
+        }
+        parent::validate();
+    }
+
+    public function toArray($noStream = false)
+    {
+        $res = [];
+        if (null !== $this->images) {
+            if (\is_array($this->images)) {
+                $res['Images'] = [];
+                $n1 = 0;
+                foreach ($this->images as $item1) {
+                    $res['Images'][$n1] = null !== $item1 ? $item1->toArray($noStream) : $item1;
+                    ++$n1;
+                }
+            }
+        }
+
+        if (null !== $this->pageNumber) {
+            $res['PageNumber'] = $this->pageNumber;
+        }
+
+        if (null !== $this->pageSize) {
+            $res['PageSize'] = $this->pageSize;
+        }
+
+        if (null !== $this->requestId) {
+            $res['RequestId'] = $this->requestId;
+        }
+
+        if (null !== $this->totalCount) {
+            $res['TotalCount'] = $this->totalCount;
+        }
+
+        return $res;
+    }
+
+    public function toMap($noStream = false)
+    {
+        return $this->toArray($noStream);
+    }
+
+    public static function fromMap($map = [])
+    {
+        $model = new self();
+        if (isset($map['Images'])) {
+            if (!empty($map['Images'])) {
+                $model->images = [];
+                $n1 = 0;
+                foreach ($map['Images'] as $item1) {
+                    $model->images[$n1] = images::fromMap($item1);
+                    ++$n1;
+                }
+            }
+        }
+
+        if (isset($map['PageNumber'])) {
+            $model->pageNumber = $map['PageNumber'];
+        }
+
+        if (isset($map['PageSize'])) {
+            $model->pageSize = $map['PageSize'];
+        }
+
+        if (isset($map['RequestId'])) {
+            $model->requestId = $map['RequestId'];
+        }
+
+        if (isset($map['TotalCount'])) {
+            $model->totalCount = $map['TotalCount'];
+        }
+
+        return $model;
+    }
+}