+++ vs-20181212/src/Models/ListRenderingImagesResponseBody/images.php
@@ -1,0 +1,90 @@
+<?php
+
+// This file is auto-generated, don't edit it. Thanks.
+
+namespace AlibabaCloud\SDK\Vs\V20181212\Models\ListRenderingImagesResponseBody;
+
+use AlibabaCloud\Dara\Model;
+
+class images extends Model
+{
+    /**
+     * @var string
+     */
+    public $description;
+
+    /**
+     * @var string
+     */
+    public $imageId;
+
+    /**
+     * @var string
+     */
+    public $name;
+
+    /**
+     * @var string
+     */
+    public $publishTime;
+    protected $_name = [
+        'description' => 'Description',
+        'imageId' => 'ImageId',
+        'name' => 'Name',
+        'publishTime' => 'PublishTime',
+    ];
+
+    public function validate()
+    {
+        parent::validate();
+    }
+
+    public function toArray($noStream = false)
+    {
+        $res = [];
+        if (null !== $this->description) {
+            $res['Description'] = $this->description;
+        }
+
+        if (null !== $this->imageId) {
+            $res['ImageId'] = $this->imageId;
+        }
+
+        if (null !== $this->name) {
+            $res['Name'] = $this->name;
+        }
+
+        if (null !== $this->publishTime) {
+            $res['PublishTime'] = $this->publishTime;
+        }
+
+        return $res;
+    }
+
+    public function toMap($noStream = false)
+    {
+        return $this->toArray($noStream);
+    }
+
+    public static function fromMap($map = [])
+    {
+        $model = new self();
+        if (isset($map['Description'])) {
+            $model->description = $map['Description'];
+        }
+
+        if (isset($map['ImageId'])) {
+            $model->imageId = $map['ImageId'];
+        }
+
+        if (isset($map['Name'])) {
+            $model->name = $map['Name'];
+        }
+
+        if (isset($map['PublishTime'])) {
+            $model->publishTime = $map['PublishTime'];
+        }
+
+        return $model;
+    }
+}