+++ vs-20181212/src/Models/UpgradeRenderingInstanceImageRequest.php
@@ -6,19 +6,19 @@ namespace AlibabaCloud\SDK\Vs\V20181212\Models;
 
 use AlibabaCloud\Dara\Model;
 
-class RebootRenderingServerRequest extends Model
+class UpgradeRenderingInstanceImageRequest extends Model
 {
     /**
-     * @var bool
+     * @var string
      */
-    public $precheck;
+    public $imageId;
 
     /**
      * @var string[]
      */
     public $renderingInstanceIds;
     protected $_name = [
-        'precheck' => 'Precheck',
+        'imageId' => 'ImageId',
         'renderingInstanceIds' => 'RenderingInstanceIds',
     ];
 
@@ -33,8 +33,8 @@ class RebootRenderingServerRequest extends Model
     public function toArray($noStream = false)
     {
         $res = [];
-        if (null !== $this->precheck) {
-            $res['Precheck'] = $this->precheck;
+        if (null !== $this->imageId) {
+            $res['ImageId'] = $this->imageId;
         }
 
         if (null !== $this->renderingInstanceIds) {
@@ -59,8 +59,8 @@ class RebootRenderingServerRequest extends Model
     public static function fromMap($map = [])
     {
         $model = new self();
-        if (isset($map['Precheck'])) {
-            $model->precheck = $map['Precheck'];
+        if (isset($map['ImageId'])) {
+            $model->imageId = $map['ImageId'];
         }
 
         if (isset($map['RenderingInstanceIds'])) {